per-cell proximal synapses.
 inspired by @fergalbyrne - but taking a slightly different approach: rather than explicitly representing an inhibitory column cell, any cells in a column having excitation within :dominance-margin of the winner cell become active. usually that will be all (bursting) or just the single winner.

the winner cell is chosen by its combined proximal and distal excitation. in the original column scheme, when the column bursts we chose the learning (winner) cell based on not-yet-connected synapses, i.e. below connected permanence threshold. now we can't do that. so instead, maintain a continuing level of excitation on active cells - 'hysteresis' - allowing one cell to be picked as the winner repeatedly until the synapses become connected. seems biologically reasonable to me.