proximal synapses on column segments, not cells. bulk learning.
proximal segments allow temporal pooling to collect predictable sequences without the sub-states interfering. keeping the proximal field at the column rather than cell level maintains generalisation over the cells in the column, leaving cells to purely represent context, and is a lot faster.

learning updates to synpases graphs are now calculated in a read-only pass, and then applied in one bulk-learn invocation. this allows the learning updates to be inspected more easily; should be faster through more use of transients; and should be straightforward to adapt to a database backend.